exam-ace-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
python-jose[cryptography]==3.3.0
asyncpg==0.30.0
redis==5.2.1
python-dotenv==1.0.1
